Finished blog hosting stuff

+++ blog_posts/generate.py
@@ -16,6 +16,10 @@ with open("templates/post-bottom.html") as f:
     post_bottom = f.read()
 
 for post_dir in post_dirs:
+
+    if "no_publish" in os.listdir("posts/%s" % post_dir):
+        continue
+
     post_fields = post_dir.split("_")
 
     year = int(post_fields[0])
@@ -24,12 +28,22 @@ for post_dir in post_dirs:
     date = datetime.datetime(year, month, day)
 
     post_name = "_".join(post_fields[3:])
+    title = post_name.replace("_", " ")
+    title = title[0].upper() + title[1:]
+    posts[date] = post_name
+
+    if os.path.exists("../site/blog-posts/%s.html" % post_name) and os.path.getmtime("posts/%s/%s.ipynb" % (post_dir, post_name)) < os.path.getmtime("../site/blog-posts/%s.html" % post_name):
+        continue
+
     command = "jupyter nbconvert --to html --template basic posts/%s/%s.ipynb" % (post_dir, post_name)
     check_call(command.split(" "))
     shutil.move("posts/%s/%s.html" % (post_dir, post_name), "tmp/")
 
     content = post_top
-    content += "<br><a href=http://awebb.info/blog-posts-py/%s.ipynb>Download the .ipynb file of this blog post here.</a><br><br>\n" % post_name
+    content += "<h1>%s</h1>\n" % title
+    date_str = date.strftime("%d %B %Y").lstrip("0")
+    content += "<p>%s</p>\n" % date_str
+    content += '<br><a href="../blog-posts-py/%s.ipynb" download target="_blank">Download the .ipynb file of this blog post here.</a><br><br>\n' % post_name
     with open("tmp/%s.html" % post_name) as f:
         content += f.read()
     content += post_bottom
@@ -37,8 +51,6 @@ for post_dir in post_dirs:
     with open("../site/blog-posts/%s.html" % post_name, "w") as f:
         f.write(content)
     shutil.copy("posts/%s/%s.ipynb" % (post_dir, post_name), "../site/blog-posts-py/")
-
-    posts[date] = post_name
 
 shutil.rmtree("tmp")
 
@@ -51,8 +63,9 @@ content = list_top
 sorted_keys = sorted(posts.keys(), reverse=True)
 month_year_str = ""
 for date in sorted_keys:
-    new_month_year_str = "\n<br><b>%s</b><br>\n" % date.strftime("%B %Y")
+    new_month_year_str = "\n<b>%s</b><br>\n" % date.strftime("%B %Y")
     if new_month_year_str != month_year_str:
+        content += "<br>"
         content += new_month_year_str
         month_year_str = new_month_year_str
     post_name = posts[date]
@@ -60,7 +73,12 @@ for date in sorted_keys:
     title = title[0].upper() + title[1:]
     date_str = date.strftime("%d/%m/%y")
     content += '%s <a href="../blog/%s">%s</a>\n' % (date_str, post_name, title)
-    content += ' <a href="../blog-posts-py/%s.ipynb"><img src="../misc/jupyter.png", style="border:0;", height="20", width="20"></a><br>\n' % post_name
+    content += '&nbsp;&nbsp;<a href="../blog-posts-py/%s.ipynb" download target="_blank">(.ipynb file)</a>\n' % post_name
+    #content += '&nbsp;<a href="../blog-posts-py/%s.ipynb">(interactive)</a>\n' % post_name
+    content += "<br>"
+
+if len(sorted_keys) == 0:
+    content += "<p><b>There are currently no blog posts here.</b></p>\n"
 
 content += list_bottom
